Add min height to Topbar component and create utility function for gap variants.
