+++ src/utilities/functions.js
@@ -11,3 +11,18 @@ export function prefix() {
  * @return  {string} - Merged TailwindCSS classes.
  */
 export const cn = ( ...classNames ) => twMerge( clsx( ...classNames ) );
+
+export const getGapClass = (gap) => {
+	const gapClasses = {
+		0: 'gap-0',
+		xxs: 'gap-1',
+		xs: 'gap-2',
+		sm: 'gap-3',
+		md: 'gap-4',
+		lg: 'gap-5',
+		xl: 'gap-6',
+		'2xl': 'gap-8',
+	};
+
+	return gapClasses[gap] || gapClasses['md']; 
+};
